cloud_run/requirements.txt
functions-framework==3.*
flask
requests
google.generativeai
google-cloud-firestore